middlewares, policy

+++ routes/api.php
@@ -1,10 +1,7 @@
 <?php
 
 use App\Http\Controllers\AccessController;
-use App\Http\Controllers\AuthController;
-use App\Http\Controllers\CartController;
 use App\Http\Controllers\FileController;
-use App\Http\Controllers\ProductController;
 use App\Http\Controllers\UserController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
@@ -29,11 +26,11 @@ Route::middleware('auth:sanctum')->group(function () {
         Route::post('/', [FileController::class, 'store']);
         Route::get('/disk', [FileController::class, 'index']);
         Route::get('/shared', [FileController::class, 'shared']);
-        Route::get('/{file}', [FileController::class, 'show']);
-        Route::patch('/{file}', [FileController::class, 'update']);
-        Route::delete('/{file}', [FileController::class, 'destroy']);
+        Route::get('/{file}', [FileController::class, 'show'])->middleware('can:download,file');
+        Route::patch('/{file}', [FileController::class, 'update'])->middleware('can:update,file');
+        Route::delete('/{file}', [FileController::class, 'destroy'])->middleware('can:delete,file');
 
-        Route::post('/{file}/accesses', [AccessController::class, 'store']);
-        Route::delete('/{file}/accesses', [AccessController::class, 'destroy']);
+        Route::post('/{file}/accesses', [AccessController::class, 'store'])->middleware('can:accessAdd,file');
+        Route::delete('/{file}/accesses', [AccessController::class, 'destroy'])->middleware('can:accessRemove,file');
     });
 });
